requirements/base.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0

channels==3.0.4
channels-redis==3.3.1

charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0

uvicorn[standard]==0.16.0

# django
Django==3.2.10
django-3-jet==1.0.7
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4

hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
msgpack==1.0.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytz==2021.3
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted[tls,http2]==21.7.0
txaio==21.2.1
typing-extensions==4.0.0
tzdata==2021.5
urllib3==1.26.7

zope.interface==5.4.0